Extract the DTD support out to its own mixin, add connect method
