--- pkgs/dart_tooling_mcp_server/lib/src/dtd.dart
@@ -1,79 +1,0 @@
-// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-// TODO: Consider moving this to a shared location, possible under the dtd
-// package.
-
-import 'dart:async';
-
-import 'package:dtd/dtd.dart';
-import 'package:json_rpc_2/json_rpc_2.dart';
-
-/// Adds the [getDebugSessions] method to [DartToolingDaemon], so that calling
-/// the Editor.getDebugSessions service method can be wrapped nicely behind a
-/// method call from a given client.
-extension GetDebugSessions on DartToolingDaemon {
-  Future<GetDebugSessionsResponse> getDebugSessions() async {
-    final result = await call(
-      'Editor',
-      'getDebugSessions',
-      params: GetDebugSessionsRequest(),
-    );
-    return GetDebugSessionsResponse.fromDTDResponse(result);
-  }
-}
-
-/// The request type for the `Editor.getDebugSessions` extension method.
-extension type GetDebugSessionsRequest.fromJson(Map<String, Object?> _value)
-    implements Map<String, Object?> {
-  factory GetDebugSessionsRequest({bool? verbose}) =>
-      GetDebugSessionsRequest.fromJson({
-        if (verbose != null) 'verbose': verbose,
-      });
-
-  bool? get verbose => _value['verbose'] as bool?;
-}
-
-/// The response type for the `Editor.getDebugSessions` extension method.
-extension type GetDebugSessionsResponse.fromJson(Map<String, Object?> _value)
-    implements Map<String, Object?> {
-  static const String type = 'GetDebugSessionsResult';
-
-  List<DebugSession> get debugSessions =>
-      (_value['debugSessions'] as List).cast<DebugSession>();
-
-  factory GetDebugSessionsResponse.fromDTDResponse(DTDResponse response) {
-    // Ensure that the response has the type you expect.
-    if (response.type != type) {
-      throw RpcException.invalidParams(
-        'Expected DTDResponse.type to be $type, got: ${response.type}',
-      );
-    }
-    return GetDebugSessionsResponse.fromJson(response.result);
-  }
-}
-
-/// An individual debug session.
-extension type DebugSession.fromJson(Map<String, Object?> _value)
-    implements Map<String, Object?> {
-  String get debuggerType => _value['debuggerType'] as String;
-  String get id => _value['id'] as String;
-  String get name => _value['name'] as String;
-  String get projectRootPath => _value['projectRootPath'] as String;
-  String get vmServiceUri => _value['vmServiceUri'] as String;
-
-  factory DebugSession({
-    required String debuggerType,
-    required String id,
-    required String name,
-    required String projectRootPath,
-    required String vmServiceUri,
-  }) => DebugSession.fromJson({
-    'debuggerType': debuggerType,
-    'id': id,
-    'name': name,
-    'projectRootPath': projectRootPath,
-    'vmServiceUri': vmServiceUri,
-  });
-}
